core/nwb.icephys.yaml: Avoid requiring 64bit integers (#342)
The author introduced 64bit unsigned integers with the introduction of
the sweep table in pynwb with 21f91920 (Add a SweepTable for icephys
data, 2018-11-02).

As it turns out this is not a clever thing to do. For one the types in
the schema are always minimum types so if we ever had 64bit
sweep numbers, this would be schema compliant as well.

And some software packages, like Igor Pro 8, can not write 64bit
integers.

So let's only require it to be uint.